refactor: move utils into separate file
