+++ src/index.ts
@@ -1,6 +1,6 @@
 import { Env } from '../worker-configuration';
 import { AppConfig } from './config';
-import { corsHeaders, createJsonResponse } from './utils';
+import { corsHeaders, createJsonResponse } from './utils/requests-responses';
 import { BnsApiDO } from './durable-objects/bns-do';
 import { HiroApiDO } from './durable-objects/hiro-api-do';
 import { StxCityDO } from './durable-objects/stx-city-do';
